fix(storage): preserve portable principal identities
Keep the live alias-to-UID directory target-portable, update all-target benchmark fixtures to stable UIDs, and satisfy Rust 1.95 Clippy without weakening the format tests.
